break out more little traits
