+++ zebra-consensus/src/checkpoint/list.rs
@@ -17,7 +17,7 @@ use std::{
 };
 
 use zebra_chain::block;
-use zebra_chain::parameters::{genesis_hash, Network};
+use zebra_chain::parameters::Network;
 
 /// The hard-coded checkpoints for mainnet, generated using the
 /// `zebra-checkpoints` tool.
@@ -42,6 +42,22 @@ const MAINNET_CHECKPOINTS: &str = include_str!("main-checkpoints.txt");
 /// See [`MAINNET_CHECKPOINTS`] for detailed `zebra-checkpoints` usage
 /// information.
 const TESTNET_CHECKPOINTS: &str = include_str!("test-checkpoints.txt");
+pub trait ParameterCheckpoint {
+    fn genesis_hash(&self) -> zebra_chain::block::Hash;
+}
+
+impl ParameterCheckpoint for Network {
+    fn genesis_hash(&self) -> zebra_chain::block::Hash {
+        match self {
+            // zcash-cli getblockhash 0
+            Network::Mainnet => "00040fe8ec8471911baa1db1266ea15dd06b4a8a5c453883c000b031973dce08",
+            // zcash-cli -testnet getblockhash 0
+            Network::Testnet => "05a60a92d99d85997cce3b87616c089f6124d7342af37106edc76126334a2c38",
+        }
+        .parse()
+        .expect("hard-coded hash parses")
+    }
+}
 
 /// A list of block height and hash checkpoints.
 ///
