+++ zebra-consensus/src/parameters/subsidy.rs
@@ -7,7 +7,7 @@ use lazy_static::lazy_static;
 use zebra_chain::{
     amount::COIN,
     block::{Height, HeightDiff},
-    parameters::Network,
+    parameters::{Network, NetworkUpgrade},
 };
 
 /// An initial period from Genesis to this Height where the block subsidy is gradually incremented. [What is slow-start mining][slow-mining]
@@ -200,13 +200,32 @@ pub const FUNDING_STREAM_ECC_ADDRESSES_MAINNET: [&str; FUNDING_STREAMS_NUM_ADDRE
 
 pub trait ParameterSubsidy {
     fn num_funding_streams(&self) -> usize;
+    fn height_for_first_halving(&self) -> Height;
 }
 
+/// Network methods related to Block Subsidy and Funding Streams
 impl ParameterSubsidy for Network {
+    /// Number of addresses for each funding stream in the Network.
+    /// [7.10]: https://zips.z.cash/protocol/protocol.pdf#fundingstreams
     fn num_funding_streams(&self) -> usize {
         match self {
             Network::Mainnet => FUNDING_STREAMS_NUM_ADDRESSES_MAINNET,
             Network::Testnet => FUNDING_STREAMS_NUM_ADDRESSES_TESTNET,
+        }
+    }
+    /// Returns the minimum height after the first halving
+    /// as described in [protocol specification §7.10][7.10]
+    ///
+    /// [7.10]: https://zips.z.cash/protocol/protocol.pdf#fundingstreams
+    fn height_for_first_halving(&self) -> Height {
+        // First halving on Mainnet is at Canopy
+        // while in Testnet is at block constant height of `1_116_000`
+        // https://zips.z.cash/protocol/protocol.pdf#zip214fundingstreams
+        match self {
+            Network::Mainnet => NetworkUpgrade::Canopy
+                .activation_height(*self)
+                .expect("canopy activation height should be available"),
+            Network::Testnet => FIRST_HALVING_TESTNET,
         }
     }
 }
