+++ zebra-rpc/src/methods/get_block_template_rpcs.rs
@@ -23,7 +23,7 @@ use zebra_chain::{
     work::difficulty::{ExpandedDifficulty, U256},
 };
 use zebra_consensus::{
-    funding_stream_address, funding_stream_values, height_for_first_halving, miner_subsidy,
+    funding_stream_address, funding_stream_values, miner_subsidy, ParameterSubsidy as _,
     RouterError,
 };
 use zebra_network::AddressBookPeers;
@@ -1098,7 +1098,7 @@ where
                 best_chain_tip_height(&latest_chain_tip)?
             };
 
-            if height < height_for_first_halving(network) {
+            if height < network.height_for_first_halving() {
                 return Err(Error {
                     code: ErrorCode::ServerError(0),
                     message: "Zebra does not support founders' reward subsidies, \
